quantity_m2 iin add order produuct => float

+++ work/server/db/models/ordersproducts.js
@@ -23,7 +23,7 @@ module.exports = (sequelize, DataTypes) => {
     {
       order_id: DataTypes.INTEGER,
       product_id: DataTypes.INTEGER,
-      quantity_m2: DataTypes.INTEGER,
+      quantity_m2: DataTypes.FLOAT,
       quantity_palet: DataTypes.INTEGER,
       quantity_real: DataTypes.INTEGER,
       price_m2: DataTypes.FLOAT,
